explicit typing to handle arm/x86 issues in type casting

+++ src/lib/getpcapheader.rs
@@ -6,7 +6,7 @@ pub struct PcapHeaderInfo {
     pub caplen: u32,
     pub origlen: u32,
     pub ts_sec: i64,
-    pub ts_usec: i32,
+    pub ts_usec: i64,
 }
 
 pub fn extract_pcap_header_info(packet: &Packet) -> PcapHeaderInfo {
@@ -14,6 +14,6 @@ pub fn extract_pcap_header_info(packet: &Packet) -> PcapHeaderInfo {
         caplen: packet.header.caplen,
         origlen: packet.header.len,
         ts_sec: packet.header.ts.tv_sec,
-        ts_usec: packet.header.ts.tv_usec,
+        ts_usec: packet.header.ts.tv_usec as i64,
     }
 }
